+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Imp.hs
@@ -5,8 +5,9 @@
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Conway.Imp (spec, conwaySpec, shelleyCertsSpec) where
+module Test.Cardano.Ledger.Conway.Imp (spec) where
 
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusContext (ContextError))
 import Cardano.Ledger.Alonzo.Rules (
@@ -17,6 +18,7 @@ import Cardano.Ledger.Alonzo.Rules (
 import Cardano.Ledger.Babbage.Rules (BabbageUtxoPredFailure, BabbageUtxowPredFailure)
 import Cardano.Ledger.Babbage.TxInfo (BabbageContextError)
 import Cardano.Ledger.BaseTypes (Inject)
+import Cardano.Ledger.Conway (ConwayEra)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Rules (
   ConwayBbodyPredFailure,
@@ -34,7 +36,6 @@ import Cardano.Ledger.Conway.Rules (
 import Cardano.Ledger.Conway.TxInfo (ConwayContextError)
 import Cardano.Ledger.Shelley.API.Mempool (ApplyTx (..))
 import Cardano.Ledger.Shelley.Rules (
-  ShelleyDelegPredFailure,
   ShelleyPoolPredFailure,
   ShelleyUtxoPredFailure,
   ShelleyUtxowPredFailure,
@@ -62,6 +63,7 @@ import Test.Cardano.Ledger.Imp.Common
 spec ::
   forall era.
   ( ConwayEraImp era
+  , EraSpecificSpec era
   , Inject (BabbageContextError era) (ContextError era)
   , Inject (ConwayContextError era) (ContextError era)
   , InjectRuleFailure "LEDGER" ConwayGovPredFailure era
@@ -71,7 +73,6 @@ spec ::
   , InjectRuleFailure "LEDGER" AlonzoUtxoPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxowPredFailure era
-  , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyPoolPredFailure era
@@ -95,9 +96,9 @@ spec ::
   Spec
 spec = do
   BabbageImp.spec @era
-  withEachEraVersion @era $ conwaySpec @era
+  withEachEraVersion @era $ conwayEraGenericSpec @era
 
-conwaySpec ::
+conwayEraGenericSpec ::
   forall era.
   ( ConwayEraImp era
   , Inject (BabbageContextError era) (ContextError era)
@@ -127,7 +128,7 @@ conwaySpec ::
   , ToExpr (Event (EraRule "BBODY" era))
   ) =>
   SpecWith (ImpInit (LedgerSpec era))
-conwaySpec = do
+conwayEraGenericSpec = do
   describe "BBODY" Bbody.spec
   describe "CERTS" Certs.spec
   describe "DELEG" Deleg.spec
@@ -142,40 +143,25 @@ conwaySpec = do
   describe "UTXOS" Utxos.spec
   describe "UTXOW" Utxow.spec
 
-shelleyCertsSpec ::
-  forall era.
+conwayEraSpecificSpec ::
   ( ConwayEraImp era
   , ShelleyEraTxCert era
-  , Inject (BabbageContextError era) (ContextError era)
   , Inject (ConwayContextError era) (ContextError era)
-  , InjectRuleFailure "LEDGER" ConwayGovPredFailure era
-  , InjectRuleFailure "LEDGER" ConwayCertsPredFailure era
-  , InjectRuleFailure "LEDGER" BabbageUtxoPredFailure era
-  , InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure era
-  , InjectRuleFailure "LEDGER" AlonzoUtxowPredFailure era
-  , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
-  , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ConwayDelegPredFailure era
-  , InjectRuleFailure "LEDGER" ConwayGovCertPredFailure era
   , InjectRuleFailure "LEDGER" ConwayLedgerPredFailure era
-  , InjectRuleFailure "LEDGER" ConwayUtxoPredFailure era
-  , InjectRuleFailure "BBODY" ConwayBbodyPredFailure era
-  , InjectRuleEvent "TICK" ConwayEpochEvent era
-  , Event (EraRule "EPOCH" era) ~ ConwayEpochEvent era
-  , Event (EraRule "NEWEPOCH" era) ~ ConwayNewEpochEvent era
-  , Event (EraRule "HARDFORK" era) ~ ConwayHardForkEvent era
-  , ApplyTx era
-  , NFData (Event (EraRule "ENACT" era))
-  , ToExpr (Event (EraRule "ENACT" era))
-  , Eq (Event (EraRule "ENACT" era))
-  , Typeable (Event (EraRule "ENACT" era))
+  , InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure era
   ) =>
-  Spec
-shelleyCertsSpec = do
-  AlonzoImp.shelleyCertsSpec @era
-  withImpInit @(LedgerSpec era) $
-    forM_ (eraProtVersions @era) $ \protVer ->
-      describe ("Certificates without deposits - " <> show protVer) $
-        modifyImpInitProtVer protVer $ do
-          describe "DELEG" Deleg.shelleyCertsSpec
-          describe "UTXO" Utxo.shelleyCertsSpec
+  SpecWith (ImpInit (LedgerSpec era))
+conwayEraSpecificSpec = do
+  describe "Conway era specific Imp spec" $
+    describe "Certificates without deposits" $ do
+      describe "DELEG" Deleg.conwayEraSpecificSpec
+      describe "LEDGER" Ledger.conwayEraSpecificSpec
+      describe "RATIFY" Ratify.conwayEraSpecificSpec
+      describe "UTXO" Utxo.conwayEraSpecificSpec
+      describe "UTXOS" Utxos.conwayEraSpecificSpec
+
+instance EraSpecificSpec ConwayEra where
+  eraSpecific =
+    AlonzoImp.alonzoEraSpecificSpec
+      >> conwayEraSpecificSpec