+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Spec.hs
@@ -38,7 +38,6 @@ import Cardano.Ledger.Plutus.Language (SLanguage (..))
 import Cardano.Ledger.Shelley.API (ApplyTx)
 import Cardano.Ledger.Shelley.LedgerState (StashedAVVMAddresses)
 import Cardano.Ledger.Shelley.Rules (
-  ShelleyDelegPredFailure,
   ShelleyPoolPredFailure,
   ShelleyUtxoPredFailure,
   ShelleyUtxowPredFailure,
@@ -54,7 +53,7 @@ import qualified Test.Cardano.Ledger.Conway.BinarySpec as Binary
 import qualified Test.Cardano.Ledger.Conway.CommitteeRatifySpec as CommitteeRatify
 import qualified Test.Cardano.Ledger.Conway.DRepRatifySpec as DRepRatify
 import qualified Test.Cardano.Ledger.Conway.Imp as Imp
-import Test.Cardano.Ledger.Conway.ImpTest (ConwayEraImp)
+import Test.Cardano.Ledger.Conway.ImpTest (ConwayEraImp, EraSpecificSpec)
 import qualified Test.Cardano.Ledger.Conway.Proposals as Proposals
 import qualified Test.Cardano.Ledger.Conway.SPORatifySpec as SPORatifySpec
 import qualified Test.Cardano.Ledger.Conway.TxInfoSpec as TxInfo
@@ -65,6 +64,7 @@ spec ::
   forall era.
   ( RuleListEra era
   , ConwayEraImp era
+  , EraSpecificSpec era
   , ApplyTx era
   , DecCBOR (TxWits era)
   , DecCBOR (TxBody era)
@@ -85,7 +85,6 @@ spec ::
   , InjectRuleFailure "LEDGER" AlonzoUtxoPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxosPredFailure era
   , InjectRuleFailure "LEDGER" AlonzoUtxowPredFailure era
-  , InjectRuleFailure "LEDGER" ShelleyDelegPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxoPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyUtxowPredFailure era
   , InjectRuleFailure "LEDGER" ShelleyPoolPredFailure era
@@ -114,7 +113,6 @@ spec =
     roundTripJsonEraSpec @era
     describe "Imp" $ do
       Imp.spec @era
-      Imp.shelleyCertsSpec @era
     describe "CostModels" $ do
       CostModelsSpec.spec @era
     describe "TxWits" $ do