+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -24,6 +24,7 @@
 module Test.Cardano.Ledger.Shelley.ImpTest (
   ImpTestM,
   LedgerSpec,
+  EraSpecificSpec (..),
   SomeSTSEvent (..),
   ImpTestState,
   ImpTestEnv (..),
@@ -298,6 +299,10 @@ instance ShelleyEraImp era => ImpSpec (LedgerSpec era) where
   -- `futurePParams` are applied and the epoch number is updated to the first epoch
   -- number of the current era
   impPrepAction = passTick
+
+class EraTest era => EraSpecificSpec era where
+  eraSpecific :: SpecWith (ImpInit (LedgerSpec era))
+  eraSpecific = pure ()
 
 data SomeSTSEvent era
   = forall (rule :: Symbol).