+++ eras/conway/impl/CHANGELOG.md
@@ -98,6 +98,7 @@
 
 ### `testlib`
 
+* Added `EraSpecificSpec ConwayEra` instance
 * Added `registerRewardAccountWithDeposit`
 * Added `regDelegToDRep`
 * Generalised the following helpers and thus changed their constraints to `ConwayEraImp`: