+++ eras/mary/impl/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### `testlib`
 
+* Added `EraSpecificSpec MaryEra` instance
 * Added `Examples` module with: `ledgerExamples`, `exampleMultiAssetValue`
 * Added `Arbitrary` instance for `TransitionConfig MaryEra`
 * Added `Era` module with `MaryEraTest` class