+++ eras/shelley/impl/CHANGELOG.md
@@ -91,6 +91,8 @@
 
 ### `testlib`
 
+* Added `EraSpecificSpec ShelleyEra` instance
+* Added `EraSpecificSpec` class
 * Removed `ShelleyEraTxCert` from `ShelleyEraImp`, so added `ShelleyEraTxCert` constraint to:
   * `registerStakeCredential`
   * `delegateStake`